Algorithms

All core algorithms (DMRG, TDVP, IDMRG) are provided by MPSKit.jl. The functions below are convenience wrappers that add progress logging, automatic JLD2 file saving, and multi-sweep truncation management.

DMRG

DynamicalCorrelators.dmrg2! — Function

dmrg2!(ψ::AbstractFiniteMPS, H, truncdims::AbstractVector; alg::DMRG2=DefaultDMRG, filename::String="default_dmrg2.jld2", verbose::Union{Bool, Integer}=true, envs=environments(ψ, H))
Add some auxiliary contents to `find_groundstate` in MPSKit.jl and rename the function as `dmrg2`.

DynamicalCorrelators.dmrg2 — Function

dmrg2(ψ, H, truncdims; kwargs...)

Non-mutating version of dmrg2!. Creates a copy of the MPS ψ before running the DMRG2 optimization.

DynamicalCorrelators.dmrg2_sweep! — Function

dmrg2_sweep!(iter, ψ, H, trscheme, ϵs; alg=DefaultDMRG, filename="default_dmrg2.jld2", verbose=true)

Perform a single two-site DMRG sweep (left-to-right then right-to-left) on ψ. Wraps MPSKit's DMRG2 algorithm with progress logging, timing, and JLD2 checkpointing.

Infinite DMRG

DynamicalCorrelators.idmrg2 — Function

idmrg2(mps, operator, alg::alg_type; verbose::Union{Bool, Integer}=true, envs = environments(mps, operator)) where {alg_type <: Union{<:IDMRG, <:IDMRG2}}
Add some auxiliary contents to `find_groundstate` in MPSKit.jl and rename the function as `idmrg2`.

Default Algorithm Configurations

DynamicalCorrelators.DefaultDMRG — Constant

DefaultDMRG

Default DMRG2 algorithm configuration for ground-state optimization. Uses Lanczos eigensolver with krylovdim=3, LAPACK SVD, and bond dimension truncation at 4096.

DynamicalCorrelators.DefaultDMRG2 — Function

DefaultDMRG2(tol, krylovdim)

Construct a DMRG2 algorithm with custom eigensolver tolerance tol and Krylov dimension krylovdim. Other parameters are the same as DefaultDMRG.

DynamicalCorrelators.DefaultTDVP — Constant

DefaultTDVP

Default single-site TDVP algorithm configuration for time evolution. Uses Lanczos integrator with krylovdim=32 for accurate time-stepping.

DynamicalCorrelators.DefaultTDVP2 — Function

DefaultTDVP2(trscheme)

Construct a two-site TDVP algorithm with the given truncation scheme trscheme. Two-site TDVP allows bond dimension growth and is used for the first few time steps.

Cluster Perturbation Theory (CPT)

DynamicalCorrelators.Perioder — Type

Perioder{P<:AbstractVector{<:Integer}, T<:AbstractArray{P}, S<:AbstractArray{P}, C<:AbstractVector{<:Tuple}}

The information needed in performing periodic procedure.

DynamicalCorrelators.CPT — Type

CPT{L<:QLattice, G<:EDSolver, P<:Perioder}

CPT method for a quantum lattice system.

DynamicalCorrelators.singleParticleGreenFunction — Function

singleParticleGreenFunction(sym::Symbol, cpt::CPT, k_path::Union{AbstractVector, ReciprocalSpace}, ω_range::Union{AbstractVector,AbstractRange}; μ::Real=0.0,  dstr::Bool=false)

The single particle Green function in k-ω space.

DynamicalCorrelators.spectrum — Function

spectrum(gfpathv::AbstractVector)

The spectrum of the single particle Green function.

DynamicalCorrelators.densityofstates — Function

Density of states